+++ app/src/routes/api/v1/dataset.router.js
@@ -19,6 +19,7 @@ const ctRegisterMicroservice = require('ct-register-microservice-node');
 const { USER_ROLES } = require('app.constants');
 const InvalidRequest = require('errors/invalidRequest.error');
 const ForbiddenRequest = require('errors/forbiddenRequest.error');
+const DatasetModel = require('models/dataset.model');
 
 const router = new Router({
     prefix: '/dataset',
@@ -282,6 +283,37 @@ class DatasetRouter {
         }
 
         try {
+            if (query.sort && (query.sort.includes('user.role') || query.sort.includes('user.name'))) {
+                logger.debug('Detected sorting by user role or name');
+                const isAdmin = ['ADMIN', 'SUPERADMIN'].includes(user && user.role);
+                if (!user || !isAdmin) {
+                    ctx.throw(403, 'Sorting by user name or role not authorized.');
+                    return;
+                }
+
+                // Reset all datasets' sorting columns
+                await DatasetModel.updateMany({}, { userRole: '', userName: '' });
+
+                // Fetch info to sort again
+                const ids = await DatasetService.getAllDatasetUserIds();
+                const users = await RelationshipsService.getUsersInfoByIds(ids);
+                await Promise.all(users.map(u => DatasetModel.updateMany(
+                    { userId: u._id },
+                    {
+                        userRole: u.role ? u.role.toLowerCase() : '',
+                        userName: u.name ? u.name.toLowerCase() : '',
+                    },
+                )));
+            }
+
+            /**
+             * We'll want to limit the maximum page size in the future
+             * However, as this will cause a production BC break, we can't enforce it just now
+             */
+            // if (query['page[size]'] && query['page[size]'] > 100) {
+            //     ctx.throw(400, 'Invalid page size');
+            // }
+
             if (Object.keys(query).find(el => el.indexOf('vocabulary[') >= 0)) {
                 ctx.query.ids = await RelationshipsService.filterByVocabularyTag(query);
                 logger.debug('Ids from vocabulary-tag', ctx.query.ids);