+++ app/test/e2e/test.constants.js
@@ -2,6 +2,7 @@
 const USERS = {
     USER: {
         id: '1a10d7c6e0a37126611fd7a5',
+        name: 'test user',
         role: 'USER',
         provider: 'local',
         email: 'user@control-tower.org',
@@ -19,6 +20,7 @@ const USERS = {
     },
     MANAGER: {
         id: '1a10d7c6e0a37126611fd7a6',
+        name: 'test manager',
         role: 'MANAGER',
         provider: 'local',
         email: 'user@control-tower.org',
@@ -36,10 +38,28 @@ const USERS = {
     },
     ADMIN: {
         id: '1a10d7c6e0a37126611fd7a7',
+        name: 'test admin',
         role: 'ADMIN',
         provider: 'local',
         email: 'user@control-tower.org',
-        name: 'John Admin',
+        extraUserData: {
+            apps: [
+                'rw',
+                'gfw',
+                'gfw-climate',
+                'prep',
+                'aqueduct',
+                'forest-atlas',
+                'data4sdgs'
+            ]
+        }
+    },
+    SUPERADMIN: {
+        id: '1a10d7c6e0a37126601fd7a6',
+        role: 'SUPERADMIN',
+        provider: 'local',
+        email: 'user@control-tower.org',
+        name: 'test super admin',
         extraUserData: {
             apps: [
                 'rw',