+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-# 24/01/2020
+# 27/01/2020
 
+- Add possibility of sorting datasets by user fields (such as name or role).
 - Fix bug patching dataset without changing dataset apps.
 
 # v1.0.0